move phase correction test case
