+++ src/xreim/index.js
@@ -1,9 +1,8 @@
 import { zeroFilling } from './zeroFilling';
 import { sortX } from './sortX';
-import { phaseCorrection } from './phaseCorrection';
+
 
 export const XReIm = {
   zeroFilling,
-  sortX,
-  phaseCorrection,
+  sortX
 };
